#117 refactor: DTO 패키지 구조 변경 및 관련 코드 수정
DTO 클래스의 패키지 경로를 일관성 있게 변경
- User 관련 DTO 클래스들을 model에서 dto 패키지로 이동
- import 경로 수정하여 새로운 패키지 구조 반영
- 코드 포맷팅 개선으로 가독성 향상
