+++ src/main/java/com/fivefeeling/memory/domain/user/dto/UserTripSummaryDTO.java
@@ -1,0 +1,11 @@
+package com.fivefeeling.memory.domain.user.dto;
+
+import com.fivefeeling.memory.domain.trip.model.TripInfoResponseDTO;
+
+public record UserTripSummaryDTO(
+        String userNickName,
+        int trips,
+        TripInfoResponseDTO recentlyTrip
+) {
+
+}
