--- src/main/java/com/fivefeeling/memory/domain/user/model/UserTripSummaryDTO.java
@@ -1,11 +1,0 @@
-package com.fivefeeling.memory.domain.user.model;
-
-import com.fivefeeling.memory.domain.trip.model.TripInfoResponseDTO;
-
-public record UserTripSummaryDTO(
-    String userNickName,
-    int trips,
-    TripInfoResponseDTO recentlyTrip
-) {
-
-}
