+++ src/main/java/com/fivefeeling/memory/global/oauth/OAuth2Service.java
@@ -1,7 +1,7 @@
 package com.fivefeeling.memory.global.oauth;
 
+import com.fivefeeling.memory.domain.user.dto.UserDTO;
 import com.fivefeeling.memory.domain.user.model.User;
-import com.fivefeeling.memory.domain.user.model.UserDTO;
 import com.fivefeeling.memory.domain.user.repository.UserRepository;
 import com.fivefeeling.memory.global.common.ResultCode;
 import com.fivefeeling.memory.global.exception.CustomException;
@@ -40,9 +40,9 @@ public class OAuth2Service implements OAuth2UserService<OAuth2UserRequest, OAuth
 
       String registrationId = userRequest.getClientRegistration().getRegistrationId();
       String userNameAttributeName = userRequest.getClientRegistration()
-          .getProviderDetails()
-          .getUserInfoEndpoint()
-          .getUserNameAttributeName();
+              .getProviderDetails()
+              .getUserInfoEndpoint()
+              .getUserNameAttributeName();
 
       Map<String, Object> attributes = oAuth2User.getAttributes();
       UserDTO userProfile = OAuthAttributes.extract(registrationId, attributes);
@@ -54,14 +54,15 @@ public class OAuth2Service implements OAuth2UserService<OAuth2UserRequest, OAuth
       List<String> roles = List.of("ROLE_USER"); // 기본 권한 설정
       String token = jwtTokenProvider.createToken(userProfile.userEmail(), roles, registrationId);
 
-      Map<String, Object> customAttribute = getCustomAttribute(registrationId, userNameAttributeName, attributes, userProfile);
+      Map<String, Object> customAttribute = getCustomAttribute(registrationId, userNameAttributeName, attributes,
+              userProfile);
       customAttribute.put("token", token);
       customAttribute.put("userId", userId);
 
       return new DefaultOAuth2User(
-          Collections.singleton(new SimpleGrantedAuthority("USER")),
-          customAttribute,
-          userNameAttributeName
+              Collections.singleton(new SimpleGrantedAuthority("USER")),
+              customAttribute,
+              userNameAttributeName
       );
     } catch (OAuth2AuthenticationException e) {
       log.error("OAuth2 인증 중 오류 발생: {}", e.getMessage());
@@ -74,10 +75,10 @@ public class OAuth2Service implements OAuth2UserService<OAuth2UserRequest, OAuth
 
 
   private Map<String, Object> getCustomAttribute(
-      String registrationId,
-      String userNameAttributeName,
-      Map<String, Object> attributes,
-      UserDTO userProfile) {
+          String registrationId,
+          String userNameAttributeName,
+          Map<String, Object> attributes,
+          UserDTO userProfile) {
     Map<String, Object> customAttribute = new ConcurrentHashMap<>();
     customAttribute.put(userNameAttributeName, attributes.get(userNameAttributeName));
     customAttribute.put("provider", registrationId);
@@ -89,7 +90,8 @@ public class OAuth2Service implements OAuth2UserService<OAuth2UserRequest, OAuth
   private User updateOrSaveUser(UserDTO userProfile) {
     try {
       // 이메일과 제공자 정보를 기준으로 사용자 검색
-      Optional<User> existingUser = userRepository.findUserByUserEmailAndProvider(userProfile.userEmail(), userProfile.provider());
+      Optional<User> existingUser = userRepository.findUserByUserEmailAndProvider(userProfile.userEmail(),
+              userProfile.provider());
 
       if (existingUser.isPresent()) {
         // 기존 사용자 업데이트
@@ -102,7 +104,8 @@ public class OAuth2Service implements OAuth2UserService<OAuth2UserRequest, OAuth
       }
     } catch (DataIntegrityViolationException e) {
       log.error("이메일 중복으로 사용자 저장 실패: {}", e.getMessage());
-      throw new OAuth2AuthenticationException(new OAuth2Error("email_already_registered", "이메일이 이미 등록되어 있습니다.", null), e.getMessage());
+      throw new OAuth2AuthenticationException(new OAuth2Error("email_already_registered", "이메일이 이미 등록되어 있습니다.", null),
+              e.getMessage());
     } catch (Exception e) {
       log.error("사용자 정보 저장 또는 업데이트 중 오류 발생: {}", e.getMessage());
       throw new CustomException(ResultCode.USER_SAVE_FAILURE, e);
