+++ src/main/java/com/fivefeeling/memory/global/oauth/OAuthAttributes.java
@@ -1,6 +1,6 @@
 package com.fivefeeling.memory.global.oauth;
 
-import com.fivefeeling.memory.domain.user.model.UserDTO;
+import com.fivefeeling.memory.domain.user.dto.UserDTO;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
@@ -8,10 +8,10 @@ import java.util.function.Function;
 
 public enum OAuthAttributes {
   GOOGLE("google", (attribute) -> new UserDTO(
-      (String) attribute.get("name"),
-      (String) attribute.get("email"),
-      "google",
-      List.of("ROLE_USER") // 기본 권한 설정
+          (String) attribute.get("name"),
+          (String) attribute.get("email"),
+          "google",
+          List.of("ROLE_USER") // 기본 권한 설정
   )),
 
   KAKAO("kakao", (attribute) -> {
@@ -19,10 +19,10 @@ public enum OAuthAttributes {
     Map<String, String> profile = (Map) account.get("profile");
 
     return new UserDTO(
-        profile.get("nickname"),
-        (String) account.get("email"),
-        "kakao",
-        List.of("ROLE_USER") // 기본 권한 설정
+            profile.get("nickname"),
+            (String) account.get("email"),
+            "kakao",
+            List.of("ROLE_USER") // 기본 권한 설정
     );
   });
 
@@ -36,10 +36,10 @@ public enum OAuthAttributes {
 
   public static UserDTO extract(String registrationId, Map<String, Object> attributes) {
     return Arrays.stream(values())
-        .filter(value -> registrationId.equals(value.registrationId))
-        .findFirst()
-        .orElseThrow(IllegalArgumentException::new)
-        .of.apply(attributes);
+            .filter(value -> registrationId.equals(value.registrationId))
+            .findFirst()
+            .orElseThrow(IllegalArgumentException::new)
+            .of.apply(attributes);
   }
 
 }
